+++ kotlin/src/main/kotlin/Main.kt
@@ -1,4 +1,4 @@
-import version_0.Lexer
+import version0.Lexer
 
 fun main() {
     // git hooks commit test
