+++ lexer/src/main/kotlin/version0/IdentifierTokenMaker.kt
@@ -1,4 +1,4 @@
-package version_0
+package version0
 
 import Position
