+++ lexer/src/main/kotlin/version0/TokenProvider.kt
@@ -1,11 +1,11 @@
-package version_1
+package version0
 
 import java.io.BufferedReader
 import java.io.InputStream
 import java.io.InputStreamReader
 
 class TokenProvider(private val input: InputStream) {
-    private val lexer = Lexer_1.getDefaultLexer()
+    private val lexer = Lexer.getDefaultLexer()
     private val reader = BufferedReader(InputStreamReader(input))
     private var hasNextStatement = true
