+++ lexer/src/main/kotlin/version1/StringTokenMaker.kt
@@ -1,4 +1,4 @@
-package version_0
+package version1
 
 import Position
