+++ parser/src/main/kotlin/Parser.kt
@@ -2,8 +2,7 @@ import builder.ASTBuilder
 import builder.AssignationASTBuilder
 import builder.DeclarationASTBuilder
 import builder.MethodASTBuilder
-import version_0.Token
-import version_0.TokenType
+import version0.Token
 
 class Parser(private val astBuilders: List<ASTBuilder<ASTNode>>) {
     fun generateAST(tokens: List<Token>): ASTNode? {
