+++ token/src/main/kotlin/version1/TokenType.kt
@@ -1,4 +1,4 @@
-package version_1
+package version1
 
 enum class TokenType {
     STRING_TYPE,
